+++ CHANGELOG.md
@@ -94,3 +94,4 @@
 - 2023-09-14: Rate chat, prompts, mod, pools, +2
 - 2023-09-15: Valid mod, session, mod, scheduler, +6
 - 2023-09-15: Valid prompt, chat, scheduler, mod, +5
+- 2023-09-16: Serde prompts, smtp, gp, session, +6